git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/rsplib2@795 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ rsplib/calcapppackets.h
@@ -2,7 +2,7 @@
  * The rsplib Prototype -- An RSerPool Implementation.
  * Copyright (C) 2005 by Thomas Dreibholz, dreibh@exp-math.uni-essen.de
  *
- * $Id: cspmonitor.c 0 2005-03-02 13:34:16Z dreibh $
+ * $Id$
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -37,18 +37,18 @@
 
 struct CalcAppMessage
 {
-   uint32_t Type;
-   uint32_t JobID;
-   double   JobSize;
-   double   Completed;
+   uint32_t           Type;
+   uint32_t           JobID;
+   unsigned long long JobSize;
+   unsigned long long Completed;
 };
 
 
 struct CalcAppCookie
 {
-   uint32_t JobID;
-   double   JobSize;
-   double   Completed;
+   uint32_t           JobID;
+   unsigned long long JobSize;
+   unsigned long long Completed;
 };
